feat(universal): land upstream pool_allocator into POSIX/Switch build
Wire src/universal/pool_allocator.cpp (69 lines) into both backends.
KISAKHACK (uintptr_t) bridge on the (unsigned int)&pool[...] free-list link
write where a pointer is type-punned through a uint inside the first 4
bytes of each free item.

+++ src/universal/pool_allocator.cpp
@@ -16,7 +16,7 @@ void __cdecl Pool_Init(char *pool, pooldata_t *pooldata, unsigned int itemSize, 
     pooldata->firstFree = pool;
 
     for (itemIndex = 0; itemIndex < itemCount - 1; ++itemIndex)
-        *(unsigned int *)&pool[itemSize * itemIndex] = (unsigned int)&pool[itemSize * (itemIndex + 1)];
+        *(unsigned int *)&pool[itemSize * itemIndex] = (unsigned int)(uintptr_t)&pool[itemSize * (itemIndex + 1)];
 
     *(unsigned int *)&pool[itemSize * itemIndex] = 0;
     pooldata->activeCount = 0;
